Consolidate API logic and utilities, refactor imports
Moved all API-related functions (buy, sell, session, login, register, portfolio, reset) into a new Scripts/API.js module and removed the now-redundant handler files. Combined spinner and notification logic into Scripts/utils.js, renaming it from notificationSystem.js. Updated dashboard.js and dropdownManager.js to import from the new API and utils modules. Cleaned up dashboard.html script imports to reflect the new structure. This refactor centralizes API and utility logic for easier maintenance and reduces code duplication.
